requirements.txt
openai 
pydantic 
python-dotenv 
instructor 
graphviz 
pyvis
eval_type_backport 
rich 
